inner parts seem to be working. making gui

+++ src/main/scala/com/ronnev/editorselection/assignment/AssignmentMeasurement.scala
@@ -1,5 +1,5 @@
 package com.ronnev.editorselection.assignment
 
 trait AssignmentMeasurement {
-    def measurement(history: java.util.List[GroupAssignment], assignment: GroupAssignment) : Int
+    def measurement(history: java.util.List[GroupAssignment], assignment: GroupAssignment) : Double
 }
